net-tcp_bbr: v2: after seeing ECN, probe on first non-CE-marked ACK
Experimental knob for faster probing, to try to improve fairness of
BBRv2 ECN-adaptive pacing.

The conjecture is that some of the unfairness effects we are seeing in
the bw_hi_down case are probably mostly from issues with the
bandwidth-probing time scale.  The BBRv2 fairness dynamics when there
are ECN/loss signals are currently built on two main properties:

(1) Multiplicative decrease after probe: this means that big flows
leave more headroom than small flows. That means that any bandwidth
share growth after the decrease is more likely to come in a higher
proportion from bandwidth previously used by big flows, rather than
small flows.

(2) Equitable probing time scale: the probing time scale of small
flows has to be at least as aggressive as for big flows; and probing
cannot be in lockstep; must be either prioritized (small flows probe
more frequently) or randomized, so that any flow is equally likely to
discover and use available bandwidth.

But with the previous pacing-based ECN response up until this commit,
we were currently not maintaining part (2). Specifically, with the
current test kernels, among flows that have the minimum cwnd of 4,
slower flows pace out packets in a way that makes them have longer
packet-timed round-trip times, and thus the 4-packet-timed-round-trip
wait between probes is longer for slower flows than it is for faster
flows. So in cases where it is possible to send faster without
creating ECN marks, the faster flows are much more likely to do that
than the slower flows.

This experiment is to start normal (additive) bw-probing as soon as
there is any ACK that suggest queues are low enough (via lack of CE
mark).

Change-Id: Ic348ecddc94965f62b12e36ac2ac705304ee1f81
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>
Signed-off-by: Mr-Morat <mtarom124@gmail.com>